refactor: 영화 상세정보 저장 로직 동시성 처리
