+++ src/main/java/com/cinefinder/movie/scheduler/MovieScheduler.java
@@ -1,7 +1,7 @@
 package com.cinefinder.movie.scheduler;
 
 import com.cinefinder.movie.service.BoxOfficeService;
-import com.cinefinder.movie.service.MovieDetailService;
+import com.cinefinder.movie.service.MovieService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -12,20 +12,14 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class MovieScheduler {
     private final BoxOfficeService boxOfficeService;
-    private final MovieDetailService movieDetailService;
+    private final MovieService movieService;
 
     @Scheduled(cron = "0 0 11 * * *")
     public void fetchMovieInfoScheduler() throws Exception {
         try {
-            movieDetailService.fetchMultiplexMovieDetails();
+            movieService.fetchMultiplexMovieDetails();
         } catch (Exception e) {
             log.error("‼️ 멀티플렉스 3사 영화 상세정보 패치 중 오류 발생", e);
-        }
-
-        try {
-            boxOfficeService.fetchDailyBoxOfficeInfo();
-        } catch (Exception e) {
-            log.error("‼️ 박스오피스 정보 패치 중 오류 발생", e);
         }
     }
 }
