+++ src/main/java/com/cinefinder/screen/service/CgvScreenScheduleServiceImpl.java
@@ -3,7 +3,7 @@ package com.cinefinder.screen.service;
 import com.cinefinder.global.exception.custom.CustomException;
 import com.cinefinder.global.util.statuscode.ApiStatus;
 import com.cinefinder.movie.mapper.MovieMapper;
-import com.cinefinder.movie.service.MovieDetailService;
+import com.cinefinder.movie.service.MovieService;
 import com.cinefinder.screen.data.dto.CinemaScheduleApiResponseDto;
 import com.cinefinder.theater.mapper.TheaterMapper;
 import com.cinefinder.theater.service.BrandService;
@@ -37,7 +37,7 @@ public class CgvScreenScheduleServiceImpl implements ScreenScheduleService {
 
     private final BrandService brandService;
     private final TheaterService theaterService;
-    private final MovieDetailService movieDetailService;
+    private final MovieService movieService;
     private final ObjectMapper objectMapper = new ObjectMapper();
     private final OkHttpClient client = new OkHttpClient();
 
@@ -132,7 +132,7 @@ public class CgvScreenScheduleServiceImpl implements ScreenScheduleService {
         List<CinemaScheduleApiResponseDto> result = new ArrayList<>();
         for (JsonNode item : scheduleList) {
             String movieCode = item.path("MovieGroupCd").asText();
-            var movie = movieDetailService.fetchMovieByBrandMovieCode(brandName, movieCode);
+            var movie = movieService.fetchMovieByBrandMovieCode(brandName, movieCode);
             if (movie == null) {
                 log.warn("{}에서 찾을 수 없는 영화 정보가 있습니다. MovieCode: {}, MovieName: {}", brandName, movieCode, item.path("MovieNmKor").asText());
                 continue;
