+++ src/main/java/com/cinefinder/screen/service/MegaScreenScheduleServiceImpl.java
@@ -4,7 +4,7 @@ import com.cinefinder.global.exception.custom.CustomException;
 import com.cinefinder.global.util.statuscode.ApiStatus;
 import com.cinefinder.movie.data.Movie;
 import com.cinefinder.movie.mapper.MovieMapper;
-import com.cinefinder.movie.service.MovieDetailService;
+import com.cinefinder.movie.service.MovieService;
 import com.cinefinder.screen.data.dto.CinemaScheduleApiResponseDto;
 import com.cinefinder.theater.mapper.TheaterMapper;
 import com.cinefinder.theater.service.BrandService;
@@ -35,7 +35,7 @@ public class MegaScreenScheduleServiceImpl implements ScreenScheduleService{
 
     private final BrandService brandService;
     private final TheaterService theaterService;
-    private final MovieDetailService movieDetailService;
+    private final MovieService movieService;
     private final RestTemplate restTemplate = new RestTemplate();
     private final ObjectMapper objectMapper = new ObjectMapper();
 
@@ -85,7 +85,7 @@ public class MegaScreenScheduleServiceImpl implements ScreenScheduleService{
         List<CinemaScheduleApiResponseDto> result = new ArrayList<>();
         for (JsonNode item : scheduleList) {
             String movieCode = item.path("rpstMovieNo").asText();
-            Movie movie = movieDetailService.fetchMovieByBrandMovieCode(brandName, movieCode);
+            Movie movie = movieService.fetchMovieByBrandMovieCode(brandName, movieCode);
             if (movie == null) {
                 log.warn("{}에서 찾을 수 없는 영화 정보가 있습니다. MovieCode: {}, MovieName: {}", brandName, movieCode, item.path("MovieNm").asText());
                 continue;
@@ -94,7 +94,7 @@ public class MegaScreenScheduleServiceImpl implements ScreenScheduleService{
             CinemaScheduleApiResponseDto dto = new CinemaScheduleApiResponseDto(
                     brandService.getBrandInfo(brandName),
                     TheaterMapper.toSimplifiedTheaterDto(theaterService.getTheaterInfo(brandName, item.path("brchNo").asText())),
-                    MovieMapper.toSimplifiedMovieDto(movieDetailService.fetchMovieByBrandMovieCode(brandName, item.path("rpstMovieNo").asText())),
+                    MovieMapper.toSimplifiedMovieDto(movieService.fetchMovieByBrandMovieCode(brandName, item.path("rpstMovieNo").asText())),
                     item.path("theabKindCd").asText(),
                     item.path("playKindNm").asText(),
                     item.path("theabNo").asText(),
